backend/requirements.txt
fastapi[standard]
sqlalchemy~=2.0.0
dotenv~=0.9.9
psycopg[binary]
pyjwt[crypto]
passlib[bcrypt]